+++ sqldump-core/src/main/java/tbrugz/sqldiff/model/adapted/TCDiff.java
@@ -1,7 +1,7 @@
 package tbrugz.sqldiff.model.adapted;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
 
 import tbrugz.sqldiff.model.ChangeType;
 import tbrugz.sqldump.dbmodel.Column;