+++ sqldump-core/src/main/java/tbrugz/sqldiff/model/package-info.java
@@ -11,5 +11,5 @@
 })
 package tbrugz.sqldiff.model;
 
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapters;
+import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapters;