+++ sqldump-core/src/main/java/tbrugz/sqldump/JAXBSchemaXMLSerializer.java
@@ -11,7 +11,7 @@ import java.io.Writer;
 import java.sql.Connection;
 import java.util.Properties;
 
-import javax.xml.bind.JAXBException;
+import jakarta.xml.bind.JAXBException;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;