+++ sqldump-core/src/main/java/tbrugz/sqldump/JSONSchemaSerializer.java
@@ -4,8 +4,8 @@ import java.io.File;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBException;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;