+++ sqldump-core/src/main/java/tbrugz/sqldump/dbmodel/SchemaModel.java
@@ -5,9 +5,9 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlType;
 
 /*
 see: http://blog.bdoughan.com/2011/06/using-jaxbs-xmlaccessortype-to.html