+++ sqldump-core/src/main/java/tbrugz/sqldump/dbmodel/package-info.java
@@ -10,4 +10,4 @@
 )*/
 package tbrugz.sqldump.dbmodel;
 
-//import javax.xml.bind.annotation.*;
+//import jakarta.xml.bind.annotation.*;