+++ sqldump-core/src/test/java/tbrugz/sqldiff/model/ConstraintDiffTest.java
@@ -2,7 +2,7 @@ package tbrugz.sqldiff.model;
 
 import java.util.Arrays;
 
-import javax.xml.bind.JAXBException;
+import jakarta.xml.bind.JAXBException;
 
 import org.junit.Assert;
 import org.junit.Before;